generative-econometric-forecasting/requirements.txt
# Generative Econometric Forecasting & Demand Planning Platform
# Production Dependencies

# =====================================
# CORE DATA SCIENCE & ANALYTICS
# =====================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# =====================================
# TIME SERIES & ECONOMETRIC MODELING
# =====================================
statsmodels>=0.14.0
prophet>=1.1.0
fredapi>=0.5.0

# =====================================
# AI & LANGCHAIN INTEGRATION
# =====================================
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
openai>=1.10.0

# =====================================
# FOUNDATION MODELS & TRANSFORMERS
# =====================================
nixtla>=0.5.0                 # TimeGPT foundation model (paid API)
statsforecast>=1.7.0          # Nixtla statistical models (free)
neuralforecast>=1.7.0         # Nixtla neural models (free)
mlforecast>=0.12.0            # Nixtla ML models (free)
chronos-forecasting>=1.0.0    # Chronos transformer models
transformers>=4.36.0          # Hugging Face transformers
torch>=2.1.0                  # PyTorch for deep learning
pytorch-lightning>=2.1.0      # Training orchestration
accelerate>=0.25.0            # Model acceleration

# =====================================
# DEEP LEARNING & SYNTHETIC DATA
# =====================================
torchvision>=0.16.0           # Computer vision components
torchaudio>=2.1.0             # Audio processing
pyro-ppl>=1.8.6               # Bayesian deep learning
gpytorch>=1.11.0              # Gaussian processes

# =====================================
# UNSTRUCTURED DATA & NLP
# =====================================
nltk>=3.8.1                   # Natural language processing
spacy>=3.7.0                  # Advanced NLP
transformers[sentencepiece]   # Tokenization support
datasets>=2.15.0              # Dataset handling
beautifulsoup4>=4.12.0        # Web scraping
requests>=2.31.0              # HTTP requests
newspaper3k>=0.2.8            # News article extraction

# =====================================
# FINANCIAL DATA & NEWS APIS
# =====================================
yfinance>=0.2.25              # Yahoo Finance data
alpha-vantage>=2.3.1          # Alpha Vantage API
newsapi-python>=0.2.7         # News API integration
feedparser>=6.0.10            # RSS feed parsing

# =====================================
# ADVANCED ANALYTICS & UNCERTAINTY
# =====================================
pymc>=5.10.0                  # Bayesian modeling
arviz>=0.17.0                 # Bayesian analysis
numpyro>=0.13.2               # JAX-based probabilistic programming
jax>=0.4.20                   # JAX for high-performance computing
jaxlib>=0.4.20                # JAX library support

# =====================================
# ENVIRONMENT & CONFIGURATION
# =====================================
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
hydra-core>=1.3.0             # Configuration management

# =====================================
# VISUALIZATION & REPORTING
# =====================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# =====================================
# UTILITIES & LOGGING
# =====================================
tqdm>=4.65.0
colorama>=0.4.6

# =====================================
# DEVELOPMENT & TESTING (Optional)
# =====================================
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# =====================================
# INSTALLATION NOTES
# =====================================
# 
# Quick Install:
# pip install -r requirements.txt
#
# For Apple Silicon (M1/M2) users:
# pip install --no-binary :all: statsmodels
# 
# For Windows users with Prophet issues:
# conda install prophet
# 
# Minimal install (demo only):
# pip install pandas numpy matplotlib seaborn scikit-learn
#